+++ examples/addons/odafc_read_dwg.py
@@ -1,15 +1,17 @@
-# Copyright (c) 2020, Manfred Moitzi
+# Copyright (c) 2020-2021, Manfred Moitzi
 # License: MIT License
 from pathlib import Path
+
+import ezdxf
 from ezdxf.addons import odafc
 
-FILE = 'colorwh.dwg'
-OUTDIR = Path('~/Desktop/outbox').expanduser()
+FILE = "colorwh.dwg"
+OUTDIR = Path("~/Desktop/outbox").expanduser()
 
-doc = odafc.readfile(OUTDIR / FILE)
+doc = odafc.readfile(ezdxf.options.test_files_path / "AutodeskSamples" / FILE)
 if doc:
     msp = doc.modelspace()
-    print(f'Filename: {doc.filename}')
-    print(f'DXF Version: {doc.dxfversion} - {doc.acad_release}')
-    print(f'Modelspace has {len(msp)} entities.')
-    doc.save()
+    print(f"Filename: {doc.filename}")
+    print(f"DXF Version: {doc.dxfversion} - {doc.acad_release}")
+    print(f"Modelspace has {len(msp)} entities.")
+    doc.saveas((OUTDIR / FILE).with_suffix(".dxf"))